# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CreditConsumeRequest": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "feature_key": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Feature Key",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "reference": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          }
        },
        "required": [
          "feature_key",
          "amount"
        ],
        "title": "CreditConsumeRequest",
        "type": "object"
      },
      "CreditEntryType": {
        "enum": [
          "grant",
          "debit",
          "refund",
          "expiry",
          "adjustment"
        ],
        "title": "CreditEntryType",
        "type": "string"
      },
      "CreditGrantRequest": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "feature_key": {
            "title": "Feature Key",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          }
        },
        "required": [
          "feature_key",
          "amount"
        ],
        "title": "CreditGrantRequest",
        "type": "object"
      },
      "CreditLedgerEntry": {
        "properties": {
          "amount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount",
            "type": "string"
          },
          "balance_after": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Balance After",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "entry_type": {
            "$ref": "#/components/schemas/CreditEntryType"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "wallet_id": {
            "format": "uuid",
            "title": "Wallet Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at",
          "wallet_id",
          "entry_type",
          "amount",
          "balance_after",
          "reason",
          "reference"
        ],
        "title": "CreditLedgerEntry",
        "type": "object"
      },
      "CreditWalletResponse": {
        "properties": {
          "balance": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Balance",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "feature_key": {
            "title": "Feature Key",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at",
          "tenant_id",
          "feature_key",
          "balance"
        ],
        "title": "CreditWalletResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProductCreate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "seed_plans": {
            "default": true,
            "title": "Seed Plans",
            "type": "boolean"
          },
          "settings": {
            "additionalProperties": true,
            "title": "Settings",
            "type": "object"
          },
          "slug": {
            "maxLength": 64,
            "minLength": 2,
            "pattern": "^[a-z0-9-]+$",
            "title": "Slug",
            "type": "string"
          },
          "tenant_type": {
            "$ref": "#/components/schemas/TenantType",
            "default": "company"
          }
        },
        "required": [
          "name",
          "slug"
        ],
        "title": "ProductCreate",
        "type": "object"
      },
      "ProductResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "settings": {
            "additionalProperties": true,
            "title": "Settings",
            "type": "object"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ProductStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at",
          "name",
          "slug",
          "description",
          "status",
          "is_active",
          "settings"
        ],
        "title": "ProductResponse",
        "type": "object"
      },
      "ProductStatus": {
        "enum": [
          "active",
          "disabled",
          "archived"
        ],
        "title": "ProductStatus",
        "type": "string"
      },
      "ProductUpdate": {
        "description": "Mutate an existing product. `settings` is shallow-merged on top\nof the existing JSONB so partial patches don't wipe unrelated keys\n\u2014 e.g. updating `auth.refresh_ttl_days` leaves `features.*` alone.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "settings": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settings"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProductStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "ProductUpdate",
        "type": "object"
      },
...
      "TenantType": {
        "description": "B2B vs B2C marker. Behavior is identical under the hood (tenant\nis the billing / audit / RBAC boundary); product UIs use this to\nrender team-aware vs single-user flows.",
        "enum": [
          "company",
          "individual"
        ],
        "title": "TenantType",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "product-platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/products": {
      "get": {
        "operationId": "list_products_api_v1_admin_products_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Platform-Admin-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Platform-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductResponse"
                  },
                  "title": "Response List Products Api V1 Admin Products Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Products",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "create_product_api_v1_admin_products_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Platform-Admin-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Platform-Admin-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Product",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/products/{slug}": {
      "patch": {
        "description": "Update fields on an existing product. `settings` is merged on top\nof the existing JSONB (keys not in the patch are preserved). Common\nuse: configure per-product auth (`{\"auth\": {\"refresh_ttl_days\": 7}}`)\nor feature flags (`{\"features\": {\"google_auto_provision\": true}}`).",
        "operationId": "update_product_api_v1_admin_products__slug__patch",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Platform-Admin-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Platform-Admin-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Product",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/credits/consume": {
      "post": {
        "operationId": "consume_api_v1_credits_consume_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Acting-Tenant-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Acting-Tenant-Slug"
            }
          },
          {
            "in": "header",
            "name": "X-Platform-Admin-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Platform-Admin-Token"
            }
          },
          {
            "in": "header",
            "name": "X-Product-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Product-Slug"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditConsumeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditWalletResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Consume",
        "tags": [
          "credits"
        ]
      }
    },
    "/api/v1/credits/grant": {
      "post": {
        "operationId": "grant_api_v1_credits_grant_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Acting-Tenant-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Acting-Tenant-Slug"
            }
          },
          {
            "in": "header",
            "name": "X-Platform-Admin-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Platform-Admin-Token"
            }
          },
          {
            "in": "header",
            "name": "X-Product-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Product-Slug"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditWalletResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Grant",
        "tags": [
          "credits"
        ]
      }
    },
    "/api/v1/credits/ledger": {
      "get": {
        "operationId": "ledger_api_v1_credits_ledger_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Acting-Tenant-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Acting-Tenant-Slug"
            }
          },
          {
            "in": "header",
            "name": "X-Platform-Admin-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Platform-Admin-Token"
            }
          },
          {
            "in": "header",
            "name": "X-Product-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Product-Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CreditLedgerEntry"
                  },
                  "title": "Response Ledger Api V1 Credits Ledger Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Ledger",
        "tags": [
          "credits"
        ]
      }
    },
    "/api/v1/credits/wallets": {
      "get": {
        "operationId": "list_wallets_api_v1_credits_wallets_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Acting-Tenant-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Acting-Tenant-Slug"
            }
          },
          {
            "in": "header",
            "name": "X-Platform-Admin-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Platform-Admin-Token"
            }
          },
          {
            "in": "header",
            "name": "X-Product-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Product-Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CreditWalletResponse"
                  },
                  "title": "Response List Wallets Api V1 Credits Wallets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Wallets",
        "tags": [
          "credits"
        ]
      }
    },
...
  }
}